# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AI Content Factory 2.0 API
  version: 0.4.0
  description: P0 API contract baseline.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/projects/{projectId}/storylines:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listProjectStorylines
      description: Project storyline forest. Nodes have recursively ordered children; roots are main storylines. Siblings sort by sort_order ascending, then id ascending. Invalid IDs return 400; missing projects return 404.
      responses:
        "200": {description: Project storyline forest., content: {application/json: {schema: {$ref: "#/components/schemas/PlotLineTreeEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
    post:
      operationId: createProjectStoryline
      description: Creates a root main storyline. Its parent_id is null, relation is root, and type is main. Invalid IDs or validation failures return 400; missing projects return 404.
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateMainPlotLineRequest"}}}}
      responses:
        "201": {description: Root main storyline created., content: {application/json: {schema: {$ref: "#/components/schemas/PlotLineEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/storylines/{storylineId}/children:
    parameters:
      - $ref: "#/components/parameters/StorylineId"
    post:
      operationId: createStorylineChild
      description: Creates a child storyline under the path parent. It inherits the parent's project_id, has relation child and type child. Invalid IDs or payloads return 400; a missing parent returns 404. The parent is not a request field, preventing cross-project parenting.
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateChildPlotLineRequest"}}}}
      responses:
        "201": {description: Child storyline created., content: {application/json: {schema: {$ref: "#/components/schemas/PlotLineEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/storylines/{storylineId}:
    parameters:
      - $ref: "#/components/parameters/StorylineId"
    patch:
      operationId: updateStoryline
      description: Updates only mutable fields. parent_id, relation, project_id, and type are immutable. expected_version is required; a mismatch returns 409 VERSION_CONFLICT. A real change increments version; an identical update returns the current storyline without another increment. Invalid IDs or payloads return 400; missing storylines return 404.
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/UpdatePlotLineRequest"}}}}
      responses:
        "200": {description: Updated storyline., content: {application/json: {schema: {$ref: "#/components/schemas/PlotLineEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  parameters:
    StorylineId:
      name: storylineId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: "Invalid request. Error codes include INVALID_PROJECT_ID and VALIDATION_ERROR."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: "Resource not found. Error code: PROJECT_NOT_FOUND."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: "Optimistic concurrency conflict. Error code: VERSION_CONFLICT."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        request_id:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    PlotLineType:
      type: string
      enum: [main, child]
    PlotLineRelation:
      type: string
      enum: [root, child]
    PlotLineStatus:
      type: string
      enum: [active, completed, archived]
    PlotLine:
      type: object
      additionalProperties: false
      required: [id, project_id, parent_id, type, relation, name, summary, start_chapter, end_chapter, status, sort_order, version, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        parent_id: {type: [string, "null"], format: uuid, description: "Null for a root main storyline; otherwise the immediate parent storyline ID."}
        type: {$ref: "#/components/schemas/PlotLineType"}
        relation: {$ref: "#/components/schemas/PlotLineRelation"}
        name: {type: string, minLength: 1, maxLength: 120}
        summary: {type: string, maxLength: 5000}
        start_chapter: {type: [integer, "null"], minimum: 1}
        end_chapter: {type: [integer, "null"], minimum: 1}
        status: {$ref: "#/components/schemas/PlotLineStatus"}
        sort_order: {type: integer, minimum: 0}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
      description: Main storylines have type main, relation root, and null parent_id. Child storylines have type child, relation child, and a parent_id in the same project. When both chapter bounds are integers, start_chapter must not exceed end_chapter.
    PlotLineTreeNode:
      type: object
      additionalProperties: false
      required: [id, project_id, parent_id, type, relation, name, summary, start_chapter, end_chapter, status, sort_order, version, created_at, updated_at, children]
      properties:
        id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        parent_id: {type: [string, "null"], format: uuid}
        type: {$ref: "#/components/schemas/PlotLineType"}
        relation: {$ref: "#/components/schemas/PlotLineRelation"}
        name: {type: string, minLength: 1, maxLength: 120}
        summary: {type: string, maxLength: 5000}
        start_chapter: {type: [integer, "null"], minimum: 1}
        end_chapter: {type: [integer, "null"], minimum: 1}
        status: {$ref: "#/components/schemas/PlotLineStatus"}
        sort_order: {type: integer, minimum: 0}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        children: {type: array, items: {$ref: "#/components/schemas/PlotLineTreeNode"}}
    CreateMainPlotLineRequest:
      type: object
      additionalProperties: false
      required: [name, summary, start_chapter, end_chapter, status, sort_order]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        summary: {type: string, maxLength: 5000}
        start_chapter: {type: [integer, "null"], minimum: 1}
        end_chapter: {type: [integer, "null"], minimum: 1}
        status: {$ref: "#/components/schemas/PlotLineStatus"}
        sort_order: {type: integer, minimum: 0}
      description: Creates a main/root storyline. When both chapter bounds are integers, start_chapter must not exceed end_chapter.
    CreateChildPlotLineRequest:
      type: object
      additionalProperties: false
      required: [name, summary, start_chapter, end_chapter, status, sort_order]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        summary: {type: string, maxLength: 5000}
        start_chapter: {type: [integer, "null"], minimum: 1}
        end_chapter: {type: [integer, "null"], minimum: 1}
        status: {$ref: "#/components/schemas/PlotLineStatus"}
        sort_order: {type: integer, minimum: 0}
      description: Creates a child storyline under the path storylineId. When both chapter bounds are integers, start_chapter must not exceed end_chapter.
    UpdatePlotLineRequest:
      type: object
      additionalProperties: false
      required: [expected_version]
      minProperties: 2
      properties:
        expected_version: {type: integer, minimum: 1}
        name: {type: string, minLength: 1, maxLength: 120}
        summary: {type: string, maxLength: 5000}
        start_chapter: {type: [integer, "null"], minimum: 1}
        end_chapter: {type: [integer, "null"], minimum: 1}
        status: {$ref: "#/components/schemas/PlotLineStatus"}
        sort_order: {type: integer, minimum: 0}
      description: parent_id, relation, project_id, and type cannot be changed. When both chapter bounds are integers, start_chapter must not exceed end_chapter.
    PlotLineEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/PlotLine"}
        request_id: {type: string}
    PlotLineTree:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/PlotLineTreeNode"}}
    PlotLineTreeEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/PlotLineTree"}
        request_id: {type: string}
...